update and deleting of employee code added

+++ src/main/java/com/crrs/repository/EmployeesRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.crrs.entity.Employees;
 
-public interface EmployeesRepository extends JpaRepository<Employees, Integer> {
+public interface EmployeesRepository extends JpaRepository<Employees, Long> {
 	
 
 }
